SNOW-1454054 - Read connection configuration from file.
