+++ src/test/java/net/snowflake/client/jdbc/ConnectionIT.java
@@ -10,7 +10,6 @@ import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertThrows;
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
 import static org.junit.Assume.assumeTrue;
@@ -73,42 +72,6 @@ public class ConnectionIT extends BaseJDBCTest {
   @Test
   public void testSimpleConnection() throws SQLException {
     Connection con = getConnection();
-    try (Statement statement = con.createStatement();
-        ResultSet resultSet = statement.executeQuery("show parameters")) {
-      assertTrue(resultSet.next());
-      assertFalse(con.isClosed());
-    }
-    con.close();
-    assertTrue(con.isClosed());
-    con.close(); // ensure no exception
-  }
-
-  @Test
-  public void testThrowExceptionIfConfigurationDoesNotExist() {
-    SnowflakeUtil.systemSetEnv("SNOWFLAKE_DEFAULT_CONNECTION_NAME", "non-existent");
-    assertThrows(SnowflakeSQLException.class, () -> SnowflakeDriver.INSTANCE.connect());
-  }
-
-  @Ignore // This test could be run only on local environment where file connection.toml is configured
-  @Test
-  public void testSimpleConnectionUsingFileConfigurationToken() throws SQLException {
-    SnowflakeUtil.systemSetEnv("SNOWFLAKE_DEFAULT_CONNECTION_NAME", "aws-oauth");
-    Connection con = SnowflakeDriver.INSTANCE.connect();
-    try (Statement statement = con.createStatement();
-        ResultSet resultSet = statement.executeQuery("show parameters")) {
-      assertTrue(resultSet.next());
-      assertFalse(con.isClosed());
-    }
-    con.close();
-    assertTrue(con.isClosed());
-    con.close(); // ensure no exception
-  }
-
-  @Ignore // This test could be run only on local environment where file connection.toml is configured
-  @Test
-  public void testSimpleConnectionUsingFileConfigurationTokenFromFile() throws SQLException {
-    SnowflakeUtil.systemSetEnv("SNOWFLAKE_DEFAULT_CONNECTION_NAME", "aws-oauth-file");
-    Connection con = SnowflakeDriver.INSTANCE.connect();
     try (Statement statement = con.createStatement();
         ResultSet resultSet = statement.executeQuery("show parameters")) {
       assertTrue(resultSet.next());
