+++ src/test/java/net/snowflake/client/jdbc/ConnectionLatestIT.java
@@ -18,6 +18,7 @@ import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotEquals;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
+import static org.junit.Assert.assertThrows;
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
 
@@ -707,6 +708,44 @@ public class ConnectionLatestIT extends BaseJDBCTest {
       return;
     }
     fail();
+  }
+
+  @Test
+  public void testThrowExceptionIfConfigurationDoesNotExist() {
+    SnowflakeUtil.systemSetEnv("SNOWFLAKE_DEFAULT_CONNECTION_NAME", "non-existent");
+    assertThrows(SnowflakeSQLException.class, () -> SnowflakeDriver.INSTANCE.connect());
+  }
+
+  @Ignore // This test could be run only on local environment where file connection.toml is
+          // configured
+  @Test
+  public void testSimpleConnectionUsingFileConfigurationToken() throws SQLException {
+    SnowflakeUtil.systemSetEnv("SNOWFLAKE_DEFAULT_CONNECTION_NAME", "aws-oauth");
+    Connection con = SnowflakeDriver.INSTANCE.connect();
+    try (Statement statement = con.createStatement();
+        ResultSet resultSet = statement.executeQuery("show parameters")) {
+      assertTrue(resultSet.next());
+      assertFalse(con.isClosed());
+    }
+    con.close();
+    assertTrue(con.isClosed());
+    con.close(); // ensure no exception
+  }
+
+  @Ignore // This test could be run only on local environment where file connection.toml is
+          // configured
+  @Test
+  public void testSimpleConnectionUsingFileConfigurationTokenFromFile() throws SQLException {
+    SnowflakeUtil.systemSetEnv("SNOWFLAKE_DEFAULT_CONNECTION_NAME", "aws-oauth-file");
+    Connection con = SnowflakeDriver.INSTANCE.connect();
+    try (Statement statement = con.createStatement();
+        ResultSet resultSet = statement.executeQuery("show parameters")) {
+      assertTrue(resultSet.next());
+      assertFalse(con.isClosed());
+    }
+    con.close();
+    assertTrue(con.isClosed());
+    con.close(); // ensure no exception
   }
 
   @Test
